requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Werkzeug==3.0.0
FastAPI==0.104.0
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.1
cryptography==41.0.7
redis==5.0.1
celery==5.3.4
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3
python-dotenv==1.0.0
PyYAML==6.0.1
pydantic==2.5.0
marshmallow==3.20.1
prometheus-client==0.19.0
python-json-logger==2.0.7
gunicorn==21.2.0
gevent==23.9.1
APScheduler==3.10.4
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0
python-multipart==0.0.6
